Fix zbuffer, use floating point

+++ src/main.rs
@@ -45,7 +45,7 @@ fn barycentric(v0: &Vector2<i32>, v1: &Vector2<i32>, v2: &Vector2<i32>, p: Vecto
     return Vector3::new(1.0 - (u.x as f32 + u.y as f32) / u.z as f32, u.y as f32 / u.z as f32, u.x as f32 / u.z as f32);
 }
 
-fn draw_triangle(v0: Vector3<f32>, v1: Vector3<f32>, v2: Vector3<f32>, color: &Color, mut canvas: &mut Canvas, zbuffer: &mut Vec<i32>, width: usize, height: usize) {
+fn draw_triangle(v0: Vector3<f32>, v1: Vector3<f32>, v2: Vector3<f32>, color: &Color, mut canvas: &mut Canvas, zbuffer: &mut Vec<f32>, width: usize, height: usize) {
     let v0i: Vector2<i32> = Vector2::new(v0.x as i32, v0.y as i32);
     let v1i: Vector2<i32> = Vector2::new(v1.x as i32, v1.y as i32);
     let v2i: Vector2<i32> = Vector2::new(v2.x as i32, v2.y as i32);
@@ -57,7 +57,7 @@ fn draw_triangle(v0: Vector3<f32>, v1: Vector3<f32>, v2: Vector3<f32>, color: &C
         for y in y_min..y_max {
             let barycentric_screen = barycentric(&v0i, &v1i, &v2i, Vector2::new(x, y));
             if barycentric_screen.x >= 0.0 && barycentric_screen.y >= 0.0 && barycentric_screen.z >= 0.0 {
-                let mut z: i32 = (barycentric_screen.x * v0.z + barycentric_screen.y * v1.z + barycentric_screen.z * v2.z) as i32;
+                let mut z: f32 = barycentric_screen.x * v0.z + barycentric_screen.y * v1.z + barycentric_screen.z * v2.z;
                 if zbuffer[x as usize + width * y as usize] < z {
                     zbuffer[x as usize + width * y as usize] = z;
                     canvas.set(x as usize, y as usize, color);
@@ -71,7 +71,7 @@ fn move_and_scale(v: Vector3<f32>, m: f32, s_x: f32, s_y: f32) -> Vector3<f32> {
     Vector3::new((v.x + m) * s_x, (v.y + m) * s_y, v.z)
 }
 
-fn render_model(model: &Vec<[Vector3<f32>; 3]>, width: usize, height: usize, mut canvas: &mut Canvas, mut zbuffer: &mut Vec<i32>) {
+fn render_model(model: &Vec<[Vector3<f32>; 3]>, width: usize, height: usize, mut canvas: &mut Canvas, mut zbuffer: &mut Vec<f32>) {
     let c = &Color {r: 255, g: 255, b: 255, a: 255 };
     let c_lines = &Color {r: 0, g: 0, b: 255, a: 255 };
     let mut triangle_count: i32 = 0;
@@ -104,7 +104,7 @@ fn main() -> Result<(), ObjError> {
     let width: usize = 800;
     let height: usize = 800;
     let mut canvas = Canvas::new(width, height, Color{r: 0, g:0, b: 0, a: 255});
-    let mut zbuffer: Vec<i32> = vec![0; width * height];
+    let mut zbuffer: Vec<f32> = vec![0.0; width * height];
     let window: Window = Window::new(&canvas);
 
     let input = BufReader::new(File::open("/Users/bjornmartens/projects/tempfromgithub/tinyrenderer/obj/african_head/african_head.obj")?);
